add different yield rate for each stakable token
add a function to update the yieldrate of a particular token. 2 internal
function handle to calculation of yield earned by a staker,
_updateOneTokenToClaim() calculate the yield earned between now and the
previous calculation for *one* token. It is used when a user change the
amount of this token staked (through stake/unstake) or when the yield
change (updateYieldRate()). The second internal function
_updateTokenToClaim() is used when a
user claim its earned token and simply loop through all staked token and
run _updateOneTokenToClaim.